Move commands in dedicated folder
