+++ lib/application/query/query.ts
@@ -1,5 +1,5 @@
 import { Request, ResponseObject, ResponseToolkit } from '@hapi/hapi';
-import { UserCommand } from '../../domain/models/UserCommand.ts';
+import { UserCommand } from '../../domain/commands/UserCommand.ts';
 import { Result } from '../../domain/models/Result.ts';
 import { executeQueryUseCase } from '../../domain/usecases/ExecuteQueryUsecase.ts';
 import { APIResponse } from '../APIResponse.ts';
